requirements.txt
absl-py==2.1.0
astunparse==1.6.3
Brotli==1.0.9
certifi==2025.4.26
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
einops==0.8.0
filelock==3.17.0
flatbuffers==24.3.25
fonttools==4.53.1
gast==0.6.0
gmpy2==2.2.1
google-pasta==0.2.0
grpcio==1.65.4
h5py==3.11.0
idna==3.7
imageio==2.34.0
Jinja2==3.1.6
keras==3.1.1
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.8.3
mdurl==0.1.2
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
ml-dtypes==0.3.2
mpmath==1.3.0
namex==0.0.8
networkx==3.4.2
numpy==1.26.4
nums-from-string==0.1.2
opencv-python==4.9.0.80
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pillow==11.1.0
protobuf==4.25.4
pycocotools==2.0.10
Pygments==2.18.0
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
rich==13.7.1
scipy==1.12.0
six==1.16.0
sympy==1.13.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow==2.16.1
termcolor==2.4.0
tifffile==2024.2.12
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tqdm==4.66.2
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.0.3
wrapt==1.16.0
yt-dlp==2025.4.30
